Tidies the homepage index.js to pull in sections as components
